add drawing diagram

+++ app/src/main/java/com/checking_sensors_app/extensions/Extension.kt
@@ -3,4 +3,4 @@ package com.checking_sensors_app.extensions
 import android.content.res.Resources
 import androidx.compose.ui.unit.dp
 
-fun pxToDp(px: Int) = (px / Resources.getSystem().displayMetrics.density).dp
+fun pxToDp(px: Float) = (px / Resources.getSystem().displayMetrics.density).dp
